Power Management: Consumption Display
display shows consumption and max consumptions of ALL regions, including unpowered

Power System prefab now enables ALL powered elements on Awake to allow for configuration, then disables all non-relevant ones on start based on floor number (or hangar)

+++ Assets/Scripts/PowerManagement/PoweredElement.cs
@@ -9,7 +9,7 @@ using UnityEngine;
 public abstract class PoweredElement : MonoBehaviour
 {
     [Tooltip("Power draw while this element is powered.")]
-    public float PowerDraw;
+    public int PowerDraw;
     [SerializeField, Tooltip("Index of the light switch corresponding with the PoweredLight. -1 if there is no switch.")]
     protected int _switchIndex = -1;
